DataScripts/requirements.txt
python-dotenv
dataclasses
dataclasses_json
dataclasses_jsonschema
pandas
seqlog
spacy
setuptools
wheel
snowflake-connector-python
aiohttp
azure-storage-blob
ndjson